packages/api/requirements.txt
# Web framework
Flask>=3.0.0

# Database and Supabase
supabase
python-dotenv
SQLAlchemy>=2.0.0
